封装问题回复-异常处理

+++ src/main/java/org/vikinc/community/dto/Comment.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Comment {
     private int id;
-    private Long parentId;
+    private int parentId;
     private Integer type;
     private String commentator;
     private String content;
